+++ lib/common/src/main/java/com/blankj/common/activity/CommonActivity.java
@@ -209,7 +209,7 @@ public abstract class CommonActivity extends BaseActivity {
 
     public void showLoading(Runnable listener) {
         if (mDialogLoading != null) {
-            dismissLoading();
+            return;
         }
         mDialogLoading = new CommonDialogLoading().init(this, listener);
         mDialogLoading.show();