+++ lib/utilcode/src/main/java/com/blankj/utilcode/util/ImageUtils.java
@@ -183,14 +183,21 @@ public final class ImageUtils {
         boolean willNotCacheDrawing = view.willNotCacheDrawing();
         view.setDrawingCacheEnabled(true);
         view.setWillNotCacheDrawing(false);
-        final Bitmap drawingCache = view.getDrawingCache();
+        Bitmap drawingCache = view.getDrawingCache();
         Bitmap bitmap;
         if (null == drawingCache) {
             view.measure(View.MeasureSpec.makeMeasureSpec(0, View.MeasureSpec.UNSPECIFIED),
                     View.MeasureSpec.makeMeasureSpec(0, View.MeasureSpec.UNSPECIFIED));
             view.layout(0, 0, view.getMeasuredWidth(), view.getMeasuredHeight());
             view.buildDrawingCache();
-            bitmap = Bitmap.createBitmap(view.getDrawingCache());
+            drawingCache = view.getDrawingCache();
+            if (drawingCache != null) {
+                bitmap = Bitmap.createBitmap(drawingCache);
+            } else {
+                bitmap = Bitmap.createBitmap(view.getMeasuredWidth(), view.getMeasuredHeight(), Bitmap.Config.ARGB_8888);
+                Canvas canvas = new Canvas(bitmap);
+                view.draw(canvas);
+            }
         } else {
             bitmap = Bitmap.createBitmap(drawingCache);
         }
@@ -279,8 +286,12 @@ public final class ImageUtils {
      */
     public static Bitmap getBitmap(final InputStream is, final int maxWidth, final int maxHeight) {
         if (is == null) return null;
-        byte[] bytes = UtilsBridge.inputStream2Bytes(is);
-        return getBitmap(bytes, 0, maxWidth, maxHeight);
+        BitmapFactory.Options options = new BitmapFactory.Options();
+        options.inJustDecodeBounds = true;
+        BitmapFactory.decodeStream(is, null, options);
+        options.inSampleSize = calculateInSampleSize(options, maxWidth, maxHeight);
+        options.inJustDecodeBounds = false;
+        return BitmapFactory.decodeStream(is, null, options);
     }
 
     /**
@@ -1578,11 +1589,11 @@ public final class ImageUtils {
      * @return {@code true}: yes<br>{@code false}: no
      */
     public static boolean isImage(final String filePath) {
-        BitmapFactory.Options options = new BitmapFactory.Options();
-        options.inJustDecodeBounds = true;
         try {
-            Bitmap bitmap = BitmapFactory.decodeFile(filePath, options);
-            return bitmap != null && options.outWidth != -1 && options.outHeight != -1;
+            BitmapFactory.Options options = new BitmapFactory.Options();
+            options.inJustDecodeBounds = true;
+            BitmapFactory.decodeFile(filePath, options);
+            return options.outWidth > 0 && options.outHeight > 0;
         } catch (Exception e) {
             return false;
         }