+++ lib/utilcode/src/main/java/com/blankj/utilcode/util/UriUtils.java
@@ -16,6 +16,9 @@ import android.text.TextUtils;
 import android.util.Log;
 
 import java.io.File;
+import java.io.FileNotFoundException;
+import java.io.IOException;
+import java.io.InputStream;
 import java.lang.reflect.Array;
 import java.lang.reflect.Method;
 
@@ -67,6 +70,18 @@ public final class UriUtils {
      * @return file
      */
     public static File uri2File(@NonNull final Uri uri) {
+        File file = uri2FileReal(uri);
+        if (file != null) return file;
+        return copyUri2Cache(uri);
+    }
+
+    /**
+     * Uri to file.
+     *
+     * @param uri The uri.
+     * @return file
+     */
+    private static File uri2FileReal(@NonNull final Uri uri) {
         Log.d("UriUtils", uri.toString());
         String authority = uri.getAuthority();
         String scheme = uri.getScheme();
@@ -144,19 +159,31 @@ public final class UriUtils {
             }// end 1_0
             else if ("com.android.providers.downloads.documents".equals(authority)) {
                 final String id = DocumentsContract.getDocumentId(uri);
-                if (!TextUtils.isEmpty(id)) {
+                if (TextUtils.isEmpty(id)) {
+                    Log.d("UriUtils", uri.toString() + " parse failed(id is null). -> 1_1");
+                    return null;
+                }
+                if (id.startsWith("raw:")) {
+                    return new File(id.substring(4));
+                }
+
+                String[] contentUriPrefixesToTry = new String[]{
+                        "content://downloads/public_downloads",
+                        "content://downloads/all_downloads",
+                        "content://downloads/my_downloads"
+                };
+
+                for (String contentUriPrefix : contentUriPrefixesToTry) {
+                    Uri contentUri = ContentUris.withAppendedId(Uri.parse(contentUriPrefix), Long.valueOf(id));
                     try {
-                        final Uri contentUri = ContentUris.withAppendedId(
-                                Uri.parse("content://downloads/public_downloads"),
-                                Long.valueOf(id)
-                        );
-                        return getFileFromUri(contentUri, "1_1");
-                    } catch (NumberFormatException e) {
-                        if (id.startsWith("raw:")) {
-                            return new File(id.substring(4));
+                        File file = getFileFromUri(contentUri, "1_1");
+                        if (file != null) {
+                            return file;
                         }
+                    } catch (Exception ignore) {
                     }
                 }
+
                 Log.d("UriUtils", uri.toString() + " parse failed. -> 1_1");
                 return null;
             }// end 1_1
@@ -214,6 +241,11 @@ public final class UriUtils {
                 File fileDir = Environment.getExternalStorageDirectory();
                 return new File(fileDir, path.substring("/QQBrowser".length(), path.length()));
             }
+        } else if ("com.huawei.hidisk.fileprovider".equals(uri.getAuthority())) {
+            String path = uri.getPath();
+            if (!TextUtils.isEmpty(path)) {
+                return new File(path.replace("/root", ""));
+            }
         }
 
         final Cursor cursor = Utils.getApp().getContentResolver().query(
@@ -240,6 +272,40 @@ public final class UriUtils {
             return null;
         } finally {
             cursor.close();
+        }
+    }
+
+    private static File copyUri2Cache(Uri uri) {
+        Log.d("UriUtils", "copyUri2Cache() called");
+        InputStream is = uri2InputStream(uri);
+        File file = new File(Utils.getApp().getCacheDir(), "" + System.currentTimeMillis());
+        UtilsBridge.writeFileFromIS(file.getAbsolutePath(), is);
+        return file;
+    }
+
+    /**
+     * uri to input stream.
+     *
+     * @param uri The uri.
+     * @return the input stream
+     */
+    public static InputStream uri2InputStream(Uri uri) {
+        StringBuilder stringBuilder = new StringBuilder();
+        InputStream is = null;
+        try {
+            is = Utils.getApp().getContentResolver().openInputStream(uri);
+            return is;
+        } catch (FileNotFoundException e) {
+            e.printStackTrace();
+            return null;
+        } finally {
+            if (is != null) {
+                try {
+                    is.close();
+                } catch (IOException e) {
+                    e.printStackTrace();
+                }
+            }
         }
     }
 }