+++ lib/utilcode/src/main/java/com/blankj/utilcode/util/UtilsActivityLifecycleImpl.java
@@ -40,9 +40,13 @@ final class UtilsActivityLifecycleImpl implements Application.ActivityLifecycleC
     private int     mConfigCount     = 0;
     private boolean mIsBackground    = false;
 
-    void init() {
+    void init(Application app) {
+        app.registerActivityLifecycleCallbacks(this);
+    }
+
+    void unInit(Application app) {
         mActivityList.clear();
-        Utils.getApp().registerActivityLifecycleCallbacks(this);
+        app.unregisterActivityLifecycleCallbacks(this);
     }
 
     Activity getTopActivity() {